# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Places API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlaceCategory = typeof PlaceCategory[keyof typeof PlaceCategory];


export const PlaceCategory = {
  restaurant: 'restaurant',
  cafe: 'cafe',
  park: 'park',
  museum: 'museum',
  beach: 'beach',
  hotel: 'hotel',
  bar: 'bar',
  attraction: 'attraction',
  other: 'other',
} as const;

export interface Place {
  id: number;
  name: string;
  city: string;
  country: string;
  /** @nullable */
  description?: string | null;
  category?: PlaceCategory;
  visitedAt: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  createdAt: string;
}

export type PlaceInputCategory = typeof PlaceInputCategory[keyof typeof PlaceInputCategory];


export const PlaceInputCategory = {
  restaurant: 'restaurant',
  cafe: 'cafe',
  park: 'park',
  museum: 'museum',
  beach: 'beach',
  hotel: 'hotel',
  bar: 'bar',
  attraction: 'attraction',
  other: 'other',
} as const;

export interface PlaceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  country: string;
  description?: string;
  category: PlaceInputCategory;
  visitedAt: string;
  photoUrl?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  lat?: number;
  lng?: number;
}

export type PlaceUpdateCategory = typeof PlaceUpdateCategory[keyof typeof PlaceUpdateCategory];


export const PlaceUpdateCategory = {
  restaurant: 'restaurant',
  cafe: 'cafe',
  park: 'park',
  museum: 'museum',
  beach: 'beach',
  hotel: 'hotel',
  bar: 'bar',
  attraction: 'attraction',
  other: 'other',
} as const;

export interface PlaceUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  city?: string;
  /** @minLength 1 */
  country?: string;
  description?: string;
  category?: PlaceUpdateCategory;
  visitedAt?: string;
  photoUrl?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  lat?: number;
  lng?: number;
}

export type StatsCategoryCounts = {[key: string]: number};

export interface Stats {
  totalPlaces: number;
  totalCountries: number;
  totalCities: number;
  categoryCounts: StatsCategoryCounts;
  /** @nullable */
  mostVisitedCity?: string | null;
  /** @nullable */
  mostVisitedCountry?: string | null;
}