# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Our Places API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: places
    description: Places operations
  - name: stats
    description: Stats operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /places:
    get:
      operationId: listPlaces
      tags: [places]
      summary: List all places
      responses:
        "200":
          description: List of places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
    post:
      operationId: createPlace
      tags: [places]
      summary: Add a new place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"

  /places/{id}:
    get:
      operationId: getPlace
      tags: [places]
      summary: Get a single place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Place details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          description: Not found
    patch:
      operationId: updatePlace
      tags: [places]
      summary: Update a place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          description: Not found
    delete:
      operationId: deletePlace
      tags: [places]
      summary: Delete a place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get visit statistics
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /places/recent:
    get:
      operationId: getRecentPlaces
      tags: [places]
      summary: Get most recently visited places
      responses:
        "200":
          description: Recent places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Place:
      type: object
      required: [id, name, city, country, visitedAt, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        country:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [restaurant, cafe, park, museum, beach, hotel, bar, attraction, other]
        visitedAt:
          type: string
          format: date
        photoUrl:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time

    PlaceInput:
      type: object
      required: [name, city, country, visitedAt, category]
      properties:
        name:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
          enum: [restaurant, cafe, park, museum, beach, hotel, bar, attraction, other]
        visitedAt:
          type: string
          format: date
        photoUrl:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        lat:
          type: number
        lng:
          type: number

    PlaceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
          enum: [restaurant, cafe, park, museum, beach, hotel, bar, attraction, other]
        visitedAt:
          type: string
          format: date
        photoUrl:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        lat:
          type: number
        lng:
          type: number

    Stats:
      type: object
      required: [totalPlaces, totalCountries, totalCities, categoryCounts]
      properties:
        totalPlaces:
          type: integer
        totalCountries:
          type: integer
        totalCities:
          type: integer
        categoryCounts:
          type: object
          additionalProperties:
            type: integer
        mostVisitedCity:
          type: ["string", "null"]
        mostVisitedCountry:
          type: ["string", "null"]