# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Places API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceCategory } from './placeCategory';

export interface Place {
  id: number;
  name: string;
  city: string;
  country: string;
  /** @nullable */
  description?: string | null;
  category?: PlaceCategory;
  visitedAt: Date;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  createdAt: Date;
}