# lib/api-zod/src/generated/types/placeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Places API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceInputCategory } from './placeInputCategory';

export interface PlaceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  country: string;
  description?: string;
  category: PlaceInputCategory;
  visitedAt: Date;
  photoUrl?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  lat?: number;
  lng?: number;
}