# lib/api-zod/src/generated/types/placeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Places API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceUpdateCategory } from './placeUpdateCategory';

export interface PlaceUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  city?: string;
  /** @minLength 1 */
  country?: string;
  description?: string;
  category?: PlaceUpdateCategory;
  visitedAt?: Date;
  photoUrl?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  lat?: number;
  lng?: number;
}