# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Places API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsCategoryCounts } from './statsCategoryCounts';

export interface Stats {
  totalPlaces: number;
  totalCountries: number;
  totalCities: number;
  categoryCounts: StatsCategoryCounts;
  /** @nullable */
  mostVisitedCity?: string | null;
  /** @nullable */
  mostVisitedCountry?: string | null;
}